add finishing touches
